Run forge fmt.

+++ contracts/cpamm.t.sol
@@ -19,7 +19,7 @@ contract CPAMMTest is Test {
         amm = new CPAMM(address(token0), address(token1));
     }
 
-    function testFuzz_sqrt(uint x) public {
+    function testFuzz_sqrt(uint256 x) public {
         unchecked {
             vm.assume(x == 0 || x * x / x == x);
         }
@@ -33,7 +33,7 @@ contract CPAMMTest is Test {
         token0.approve(address(amm), 100);
         token1.approve(address(amm), 10);
 
-        uint shares0 = amm.addLiquidity(100, 10);
+        uint256 shares0 = amm.addLiquidity(100, 10);
         assertEq(shares0, 31); // sqrt(100 * 10)
 
         vm.startPrank(bob);
@@ -42,23 +42,23 @@ contract CPAMMTest is Test {
         token0.approve(address(amm), 100);
         token1.approve(address(amm), 10);
 
-        uint shares1 = amm.addLiquidity(100, 10);
+        uint256 shares1 = amm.addLiquidity(100, 10);
         assertEq(shares1, 31);
 
-        (uint amount0_0, uint amount1_0) = amm.removeLiquidity(shares1);
+        (uint256 amount0_0, uint256 amount1_0) = amm.removeLiquidity(shares1);
         assertEq(amount0_0, 100);
         assertEq(amount1_0, 10);
 
         vm.startPrank(alice);
-        (uint amount0_1, uint amount1_1) = amm.removeLiquidity(shares0);
+        (uint256 amount0_1, uint256 amount1_1) = amm.removeLiquidity(shares0);
         assertEq(amount0_1, 100);
         assertEq(amount1_1, 10);
     }
 
     function test_swap() public {
-        uint rate = 10;
-        uint supply0 = 10000000000000000;
-        uint supply1 = supply0 * rate;
+        uint256 rate = 10;
+        uint256 supply0 = 10000000000000000;
+        uint256 supply1 = supply0 * rate;
         token0.mint(supply0);
         token1.mint(supply1);
         token0.approve(address(amm), supply0);
@@ -68,7 +68,7 @@ contract CPAMMTest is Test {
         vm.startPrank(alice);
         token0.mint(10000);
         token0.approve(address(amm), 10000);
-        uint amount = amm.swap(address(token0), 10000);
+        uint256 amount = amm.swap(address(token0), 10000);
         assertEq(amount, 99699);
         assertEq(token0.balanceOf(alice), 0);
         assertEq(token1.balanceOf(alice), amount);
